requirements.txt
openai
python-dotenv
httpx
typer
beautifulsoup4
